refactor: split EventPropertyMetadataBuilder out, move generic reflection helpers to TypeUtils
- Extracted EventPropertyMetadataBuilder from EventMetadataBuilder: reflecting
  an event type's own properties/partition keys is now independent from
  resolving its topic attribute and computed fields.
- Moved FindAttributeByName, GetPropertyValue<T>, and
  MapConstructorParametersToProperties into TypeUtils, since they carry no
  event-metadata-specific meaning.
- Removed a private kebab-case fallback in EventMetadataBuilder that
  duplicated the Momentum.Extensions.Abstractions.Extensions.ToKebabCase()
  extension already imported in the same file.
- Added an all-properties-showcase integration scenario (real TestEvents
  fixture + XML doc input) as a durable, reviewable example of every
  renderable EventMetadata property, alongside the existing marker-value
  completeness unit test.

All 682 tests across the touched projects pass unchanged (byte-identical
generated markdown baselines), confirming these are behavior-preserving
refactors.
