+++ libs/Momentum/src/Momentum.Extensions.EventMarkdownGenerator/Services/EventMetadataBuilder.cs
@@ -10,6 +10,7 @@ namespace Momentum.Extensions.EventMarkdownGenerator.Services;
 ///     Builds <see cref="EventMetadata"/> for a single discovered event type. Separated from
 ///     <see cref="AssemblyEventDiscovery"/> so assembly/type scanning stays independent of the
 ///     (considerably larger) job of reflecting an event type and its topic attribute into metadata.
+///     Generic reflection helpers with no metadata-specific meaning live in <see cref="TypeUtils"/>.
 /// </summary>
 public static class EventMetadataBuilder
 {
@@ -20,15 +21,13 @@ public static class EventMetadataBuilder
         var topicAttribute = GetEventTopicAttributeDynamic(eventType, attributeNamePrefix);
         var obsoleteAttribute = eventType.GetCustomAttribute<ObsoleteAttribute>();
         var (properties, partitionKeys) =
-            GetEventPropertiesAndPartitionKeys(eventType, xmlParser, calculator, partitionKeyAttributeNamePrefix);
+            EventPropertyMetadataBuilder.Build(eventType, xmlParser, calculator, partitionKeyAttributeNamePrefix);
 
         var topicName = GetTopicName(topicAttribute, eventType);
 
         // Access properties via reflection for cross-assembly compatibility
         var (shouldPluralize, domain, isInternal, version, eventNameOverride) = GetTopicAttributeProperties(topicAttribute);
 
-        // Simple pluralization fallback - add 's' to the end
-        // This is a fallback when the extension method is not available
         if (shouldPluralize && !topicName.EndsWith("s", StringComparison.OrdinalIgnoreCase))
         {
             topicName = topicName.Pluralize();
@@ -89,8 +88,7 @@ public static class EventMetadataBuilder
 
     private static Attribute GetEventTopicAttributeDynamic(Type type, string attributeNamePrefix)
     {
-        var foundAttribute = type.GetCustomAttributes()
-            .FirstOrDefault(attr => attr.GetType().Name.StartsWith(attributeNamePrefix));
+        var foundAttribute = TypeUtils.FindAttributeByName(type.GetCustomAttributes(), attributeNamePrefix);
 
         if (foundAttribute == null)
         {
@@ -98,94 +96,6 @@ public static class EventMetadataBuilder
         }
 
         return foundAttribute;
-    }
-
-    private static (List<EventPropertyMetadata> properties, List<PartitionKeyMetadata> partitionKeys) GetEventPropertiesAndPartitionKeys(
-        Type eventType, XmlDocumentationParser? xmlParser, PayloadSizeCalculator calculator, string partitionKeyAttributeNamePrefix)
-    {
-        var properties = new List<EventPropertyMetadata>();
-        var partitionKeys = new List<PartitionKeyMetadata>();
-
-        var constructor = eventType.GetConstructors().FirstOrDefault();
-        var constructorParameters = constructor?.GetParameters() ?? [];
-
-        var parameterToPropertyMap = MapConstructorParametersToProperties(eventType, constructorParameters);
-
-        var eventDoc = xmlParser?.GetEventDocumentation(eventType);
-
-        foreach (var property in eventType.GetProperties(BindingFlags.Public | BindingFlags.Instance))
-        {
-            var isComplexType = !TypeUtils.IsPrimitiveType(property.PropertyType);
-            var isRequired = TypeUtils.IsRequiredProperty(property);
-
-            // Check for the partition key attribute by name (generic, works across assembly load contexts)
-            var partitionKeyAttr = FindAttributeByName(property.GetCustomAttributes(), partitionKeyAttributeNamePrefix);
-            var isPartitionKey = partitionKeyAttr != null;
-
-            // If not found on property, check corresponding constructor parameter for records
-            if (!isPartitionKey && parameterToPropertyMap.TryGetValue(property.Name, out var parameter))
-            {
-                partitionKeyAttr = FindAttributeByName(parameter.GetCustomAttributes(), partitionKeyAttributeNamePrefix);
-                isPartitionKey = partitionKeyAttr != null;
-            }
-
-            var partitionKeyOrder = partitionKeyAttr is null
-                ? (int?)null
-                : GetPropertyValue<int?>(partitionKeyAttr.GetType(), partitionKeyAttr, "Order") ?? 0;
-
-            var description = eventDoc?.PropertyDescriptions?.GetValueOrDefault(property.Name) ?? "No description available";
-            var sizeResult = calculator.CalculatePropertySize(property, property.PropertyType);
-
-            properties.Add(new EventPropertyMetadata
-            {
-                Name = property.Name,
-                TypeName = TypeUtils.GetFriendlyTypeName(property.PropertyType),
-                PropertyType = property.PropertyType,
-                IsRequired = isRequired,
-                IsComplexType = isComplexType,
-                IsPartitionKey = isPartitionKey,
-                PartitionKeyOrder = partitionKeyOrder,
-                Description = description,
-                EstimatedSizeBytes = sizeResult.SizeBytes,
-                IsAccurate = sizeResult.IsAccurate,
-                SizeWarning = sizeResult.Warning
-            });
-
-            if (isPartitionKey)
-            {
-                partitionKeys.Add(new PartitionKeyMetadata
-                {
-                    Name = property.Name,
-                    TypeName = TypeUtils.GetFriendlyTypeName(property.PropertyType),
-                    Description = description,
-                    Order = partitionKeyOrder ?? 0,
-                    IsFromParameter = parameterToPropertyMap.ContainsKey(property.Name)
-                });
-            }
-        }
-
-        partitionKeys = partitionKeys.OrderBy(pk => pk.Order).ThenBy(pk => pk.Name).ToList();
-
-        return (properties, partitionKeys);
-    }
-
-    private static Dictionary<string, ParameterInfo> MapConstructorParametersToProperties(Type eventType,
-        ParameterInfo[] constructorParameters)
-    {
-        var map = new Dictionary<string, ParameterInfo>(StringComparer.OrdinalIgnoreCase);
-
-        foreach (var parameter in constructorParameters)
-        {
-            var property = eventType.GetProperties(BindingFlags.Public | BindingFlags.Instance)
-                .FirstOrDefault(p => string.Equals(p.Name, parameter.Name, StringComparison.OrdinalIgnoreCase));
-
-            if (property != null)
-            {
-                map[property.Name] = parameter;
-            }
-        }
-
-        return map;
     }
 
     private static string? GetDomainFromNamespace(string? namespaceName)
@@ -208,23 +118,6 @@ public static class EventMetadataBuilder
     }
 
     /// <summary>
-    ///     Simple kebab case conversion when extension method is not available
-    /// </summary>
-    private static string ConvertToKebabCase(string input)
-    {
-        if (string.IsNullOrEmpty(input))
-            return input;
-
-        var result = string.Concat(
-            input.Select((x, i) => i > 0 && char.IsUpper(x)
-                ? "-" + char.ToLower(x)
-                : char.ToLower(x).ToString())
-        );
-
-        return result;
-    }
-
-    /// <summary>
     ///     Extracts properties from EventTopicAttribute using reflection for cross-assembly compatibility.
     /// </summary>
     private static (bool shouldPluralize, string? domain, bool isInternal, string version, string? eventName)
@@ -232,35 +125,13 @@ public static class EventMetadataBuilder
     {
         var attrType = topicAttribute.GetType();
 
-        var shouldPluralize = GetPropertyValue<bool?>(attrType, topicAttribute, "ShouldPluralizeTopicName") ?? false;
-        var domain = GetPropertyValue<string?>(attrType, topicAttribute, "Domain");
-        var isInternal = GetPropertyValue<bool?>(attrType, topicAttribute, "Internal") ?? false;
-        var version = GetPropertyValue<string?>(attrType, topicAttribute, "Version") ?? "v1";
-        var eventName = GetPropertyValue<string?>(attrType, topicAttribute, "EventName");
+        var shouldPluralize = TypeUtils.GetPropertyValue<bool?>(attrType, topicAttribute, "ShouldPluralizeTopicName") ?? false;
+        var domain = TypeUtils.GetPropertyValue<string?>(attrType, topicAttribute, "Domain");
+        var isInternal = TypeUtils.GetPropertyValue<bool?>(attrType, topicAttribute, "Internal") ?? false;
+        var version = TypeUtils.GetPropertyValue<string?>(attrType, topicAttribute, "Version") ?? "v1";
+        var eventName = TypeUtils.GetPropertyValue<string?>(attrType, topicAttribute, "EventName");
 
         return (shouldPluralize, domain, isInternal, version, eventName);
-    }
-
-    /// <summary>
-    ///     Finds a custom attribute by name (or name prefix) so discovery works across assembly load contexts
-    ///     and for any attribute type matching the configured convention, not just a hardcoded compiled type.
-    /// </summary>
-    private static Attribute? FindAttributeByName(IEnumerable<Attribute> attributes, string attributeNamePrefix) =>
-        attributes.FirstOrDefault(attr => attr.GetType().Name.StartsWith(attributeNamePrefix));
-
-    /// <summary>
-    ///     Gets a property value from an object using reflection with safe null handling.
-    /// </summary>
-    private static T? GetPropertyValue<T>(Type type, object instance, string propertyName)
-    {
-        var property = type.GetProperty(propertyName);
-
-        if (property == null)
-            return default;
-
-        var value = property.GetValue(instance);
-
-        return value is T typedValue ? typedValue : default;
     }
 
     /// <summary>
@@ -269,12 +140,8 @@ public static class EventMetadataBuilder
     private static string GetTopicName(object topicAttribute, Type eventType)
     {
         var attrType = topicAttribute.GetType();
-        var topic = GetPropertyValue<string?>(attrType, topicAttribute, "Topic");
+        var topic = TypeUtils.GetPropertyValue<string?>(attrType, topicAttribute, "Topic");
 
-        if (!string.IsNullOrEmpty(topic))
-            return topic;
-
-        // Fallback kebab case conversion when extension method is not available
-        return ConvertToKebabCase(eventType.Name);
+        return !string.IsNullOrEmpty(topic) ? topic : eventType.Name.ToKebabCase();
     }
 }
