+++ libs/Momentum/tests/Momentum.Extensions.EventMarkdownGenerator.Tests/IntegrationTestScenarios/all-properties-showcase/expected/schemas/TestEvents.AppDomain.Cashiers.Contracts.IntegrationEvents.Cashier.md
@@ -1,0 +1,27 @@
+---
+editLink: false
+---
+
+<!-- // @formatter:off -->
+<!-- prettier-ignore -->
+
+# Cashier
+
+## Description
+
+Represents a cashier entity.
+
+## Schema
+
+<!-- #region schema -->
+
+| Property | Type | Required | Description |
+| -------- | ---- | -------- | ----------- |
+| Id| `Guid` | ✓| Gets or sets the id. |
+| Name| `string` | ✓| Gets or sets the name. |
+| Email| `string` | | Gets or sets the email. |
+| IsActive| `bool` | ✓| Gets or sets the isactive. |
+
+
+<!-- #endregion schema -->
+
