+++ libs/Momentum/tests/Momentum.Extensions.EventMarkdownGenerator.Tests/IntegrationTestScenarios/all-properties-showcase/expected/schemas/TestEvents.AppDomain.Comprehensive.Contracts.IntegrationEvents.ShowcaseLocation.md
@@ -1,0 +1,25 @@
+---
+editLink: false
+---
+
+<!-- // @formatter:off -->
+<!-- prettier-ignore -->
+
+# ShowcaseLocation
+
+## Description
+
+Represents a showcaselocation entity.
+
+## Schema
+
+<!-- #region schema -->
+
+| Property | Type | Required | Description |
+| -------- | ---- | -------- | ----------- |
+| Id| `Guid` | ✓| Gets or sets the id. |
+| Name| `string` | ✓| Gets or sets the name. |
+
+
+<!-- #endregion schema -->
+
